requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
reportlab==4.2.5
django-cors-headers==4.7.0
Pillow==11.1.0
